完善文件结构
